// bot/src/types/user-data.type.ts
export interface UserData {
  chatId: number
  username?: string | null
  first_name?: string | null
  last_name?: string | null
  language_code?: string | null
  state?: string | null
  nickname?: string | null
  city?: string | null
  description?: string | null
  photoURL?: string | null
  roleId?: string | null
  recommendationArray?: number[]
  lastViewedId?: number | null
}